+++ libs/zhi-cmd/CHANGELOG.md
@@ -1,5 +1,12 @@
 # zhi-cmd
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.11.0
+
 ## 0.8.0
 
 ### Minor Changes